Move processor contract to processor package
