+++ scabbard-processor/src/main/java/dev/arunkumar/scabbard/plugin/processor/graphviz/DefaultBindingGraphProcessor.kt
@@ -5,11 +5,11 @@ import dagger.model.BindingGraph
 import dagger.model.BindingGraph.MaybeBinding
 import dagger.model.ComponentPath
 import dev.arunkumar.dot.DotGraph
-import dev.arunkumar.scabbard.plugin.BindingGraphProcessor
 import dev.arunkumar.scabbard.plugin.di.ProcessorScope
 import dev.arunkumar.scabbard.plugin.options.ScabbardOptions
 import dev.arunkumar.scabbard.plugin.output.OutputWriter
 import dev.arunkumar.scabbard.plugin.parser.subcomponentsOf
+import dev.arunkumar.scabbard.plugin.processor.BindingGraphProcessor
 import dev.arunkumar.scabbard.plugin.processor.graphviz.renderer.DaggerComponent
 import dev.arunkumar.scabbard.plugin.processor.graphviz.renderer.InheritedBinding
 import dev.arunkumar.scabbard.plugin.util.processingBlock
@@ -20,13 +20,12 @@ import kotlin.collections.component2
 
 @Suppress("UnstableApiUsage")
 @ProcessorScope
-@JvmSuppressWildcards
 class DefaultBindingGraphProcessor
 @Inject
 constructor(
   override val bindingGraph: BindingGraph,
   private val scabbardOptions: ScabbardOptions,
-  private val outputWriters: Set<OutputWriter>,
+  private val outputWriters: Set<@JvmSuppressWildcards OutputWriter>,
   private val renderingContext: RenderingContext
 ) : BindingGraphProcessor {
 
@@ -35,7 +34,6 @@ constructor(
 
     // Group all the nodes by their component
     network.nodes()
-      .asSequence()
       .groupBy(BindingGraph.Node::componentPath)
       .map { (componentPath, nodes) ->
         val currentComponent = componentPath.currentComponent()
