+++ scabbard-processor/src/main/java/dev/arunkumar/scabbard/plugin/processor/graphviz/GraphvizVisualizationModule.kt
@@ -3,10 +3,10 @@ package dev.arunkumar.scabbard.plugin.processor.graphviz
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoSet
-import dev.arunkumar.scabbard.plugin.BindingGraphProcessor
+import dev.arunkumar.scabbard.plugin.processor.BindingGraphProcessor
 
 @Module
-interface BindingGraphProcessorModule {
+interface GraphvizVisualizationModule {
   @Binds
   @IntoSet
   fun DefaultBindingGraphProcessor.bindingGraphProcessor(): BindingGraphProcessor
